pyproject.toml
[project]
name = "challenge"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2.4",
    "django-cors-headers>=4.7.0",
    "djangorestframework>=3.16.0",
    "channels>=4.0.0",
    "channels-redis>=4.1.0",
    "celery>=5.3.0",
    "redis>=5.0.0",
    "requests>=2.31.0",
    "pillow>=10.0.0",
    "python-multipart>=0.0.6",
]
